pyproject.toml
[project]
name = "voiceobs"
version = "0.0.2"
description = "Open, vendor-neutral observability for voice AI conversations using OpenTelemetry"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "voiceobs contributors" }
]
keywords = ["voice", "ai", "observability", "opentelemetry", "tracing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "typer>=0.9.0",
    "pyyaml>=6.0.0",
    "livekit-plugins-elevenlabs>=1.3.12",
    "colorlog>=6.10.1",
]

[project.optional-dependencies]
eval = [
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
]
otlp = [
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]
server = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",
    "openai>=1.0.0",
    "mutagen>=1.47.0",
    "deepgram-sdk>=3.0.0",
    "elevenlabs>=1.0.0",
    "python-dotenv>=1.0.0",
    "livekit>=1.0.23",
    "python-jose[cryptography]>=3.3.0",
]
livekit = [
    "livekit-agents[silero,turn-detector]>=1.3.11",
    "livekit-plugins-deepgram>=1.0.0",
    "livekit-plugins-openai>=1.0.0",
]
s3 = [
    "boto3>=1.28.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "pre-commit>=3.0.0",
    # Server dependencies needed for server tests
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",
    "openai>=1.0.0",
    "mutagen>=1.47.0",
    "deepgram-sdk>=3.0.0",
    "elevenlabs>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.scripts]
voiceobs = "voiceobs.cli:app"

[project.urls]
Homepage = "https://github.com/voice-observation/voiceobs"
Documentation = "https://github.com/voice-observation/voiceobs#readme"
Repository = "https://github.com/voice-observation/voiceobs"
Issues = "https://github.com/voice-observation/voiceobs/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voiceobs"]

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = ["examples", "integrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/voiceobs/server/prompts/persona.py" = ["E501"]  # Long prompt template strings

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
